Fix mobile unsupported types: - TransactionQuery
