+++ zcncore/transaction_mobile.go
@@ -15,6 +15,7 @@ import (
 	"github.com/0chain/errors"
 	"github.com/0chain/gosdk/core/block"
 	"github.com/0chain/gosdk/core/encryption"
+	"github.com/0chain/gosdk/core/resty"
 	"github.com/0chain/gosdk/core/transaction"
 	"github.com/0chain/gosdk/core/util"
 	"github.com/0chain/gosdk/core/zcncrypto"
@@ -976,7 +977,18 @@ func (b *Block) ForEachTxns(tf IterTxnFunc) {
 
 func toMobileBlock(b *block.Block) *Block {
 	lb := &Block{
-		header:            b.Header,
+		header: &BlockHeader{
+			Version:               b.Header.Version,
+			CreationDate:          b.Header.CreationDate,
+			Hash:                  b.Header.Hash,
+			MinerID:               b.Header.MinerID,
+			Round:                 b.Header.Round,
+			RoundRandomSeed:       b.Header.RoundRandomSeed,
+			MerkleTreeRoot:        b.Header.MerkleTreeRoot,
+			StateHash:             b.Header.StateHash,
+			ReceiptMerkleTreeRoot: b.Header.ReceiptMerkleTreeRoot,
+			NumTxns:               b.Header.NumTxns,
+		},
 		MinerID:           string(b.MinerID),
 		Round:             b.Round,
 		RoundRandomSeed:   b.RoundRandomSeed,
@@ -1097,7 +1109,7 @@ func GetBlockByRound(numSharders int, round int64, tm *ReqTimeout) (b *Block, er
 		}
 
 		b = toMobileBlock(respo.Block)
-		b.Header = respo.Header
+		b.header = respo.Header
 
 		var h = encryption.FastHash([]byte(b.Hash))
 		if roundConsensus[h]++; roundConsensus[h] > maxConsensus {
@@ -1162,4 +1174,152 @@ func GetMagicBlockByNumber(numSharders int, number int64, tm *ReqTimeout) (m *bl
 	}
 
 	return
+}
+
+// FromAll query transaction from all sharders whatever it is selected or offline in previous queires, and return consensus result
+func (tq *TransactionQuery) FromAll(query string, handle QueryResultHandle, tm *ReqTimeout) error {
+	if tq == nil || tq.max == 0 {
+		return ErrNoAvailableSharders
+	}
+
+	ctx, cancel := makeTimeoutContext(tm)
+	defer cancel()
+
+	urls := make([]string, 0, tq.max)
+	for _, host := range tq.sharders {
+		urls = append(urls, tq.buildUrl(host, query))
+	}
+
+	r := resty.New()
+	r.DoGet(ctx, urls...).
+		Then(func(req *http.Request, resp *http.Response, respBody []byte, cf context.CancelFunc, err error) error {
+			res := QueryResult{
+				Content:    respBody,
+				Error:      err,
+				StatusCode: http.StatusBadRequest,
+			}
+
+			if resp != nil {
+				res.StatusCode = resp.StatusCode
+
+				Logger.Debug(req.URL.String() + " " + resp.Status)
+				Logger.Debug(string(respBody))
+			} else {
+				Logger.Debug(req.URL.String())
+
+			}
+
+			if handle != nil {
+				if handle(res) {
+
+					cf()
+				}
+			}
+
+			return nil
+		})
+
+	r.Wait()
+
+	return nil
+}
+
+// FromAny query transaction from any sharder that is not selected in previous queires. use any used sharder if there is not any unused sharder
+func (tq *TransactionQuery) FromAny(query string, tm *ReqTimeout) (QueryResult, error) {
+	res := QueryResult{
+		StatusCode: http.StatusBadRequest,
+	}
+
+	ctx, cancel := makeTimeoutContext(tm)
+	defer cancel()
+
+	err := tq.validate(1)
+
+	if err != nil {
+		return res, err
+	}
+
+	host, err := tq.randOne(ctx)
+
+	if err != nil {
+		return res, err
+	}
+
+	r := resty.New()
+	requestUrl := tq.buildUrl(host, query)
+
+	Logger.Debug("GET", requestUrl)
+
+	r.DoGet(ctx, requestUrl).
+		Then(func(req *http.Request, resp *http.Response, respBody []byte, cf context.CancelFunc, err error) error {
+			res.Error = err
+			if err != nil {
+				return err
+			}
+
+			res.Content = respBody
+			Logger.Debug(string(respBody))
+
+			if resp != nil {
+				res.StatusCode = resp.StatusCode
+			}
+
+			return nil
+		})
+
+	errs := r.Wait()
+
+	if len(errs) > 0 {
+		return res, errs[0]
+	}
+
+	return res, nil
+
+}
+
+func (tq *TransactionQuery) GetInfo(query string, tm *ReqTimeout) (*QueryResult, error) {
+
+	consensuses := make(map[int]int)
+	var maxConsensus int
+	var consensusesResp QueryResult
+	// {host}{query}
+
+	ctx, cancel := makeTimeoutContext(tm)
+	defer cancel()
+
+	err := tq.FromAll(ctx, query,
+		func(qr QueryResult) bool {
+			//ignore response if it is network error
+			if qr.StatusCode >= 500 {
+				return false
+			}
+
+			consensuses[qr.StatusCode]++
+			if consensuses[qr.StatusCode] >= maxConsensus {
+				maxConsensus = consensuses[qr.StatusCode]
+				consensusesResp = qr
+			}
+
+			return false
+
+		})
+
+	if err != nil {
+		return nil, err
+	}
+
+	if maxConsensus == 0 {
+		return nil, errors.New("zcn: query not found")
+	}
+
+	rate := float32(maxConsensus*100) / float32(tq.max)
+	if rate < consensusThresh {
+		return nil, ErrInvalidConsensus
+	}
+
+	if consensusesResp.StatusCode != http.StatusOK {
+		return nil, errors.New(string(consensusesResp.Content))
+	}
+
+	return &consensusesResp, nil
 }
