+++ zcncore/transaction_query.go
@@ -63,6 +63,11 @@ func NewTransactionQuery(sharders []string) (*TransactionQuery, error) {
 	return tq, nil
 }
 
+func (tq *TransactionQuery) Reset() {
+	tq.selected = make(map[string]interface{})
+	tq.offline = make(map[string]interface{})
+}
+
 // validate validate data and input
 func (tq *TransactionQuery) validate(num int) error {
 	if tq == nil || tq.max == 0 {
@@ -83,11 +88,6 @@ func (tq *TransactionQuery) validate(num int) error {
 
 	return nil
 
-}
-
-func (tq *TransactionQuery) Reset() {
-	tq.selected = make(map[string]interface{})
-	tq.offline = make(map[string]interface{})
 }
 
 // buildUrl build url with host and parts
@@ -179,151 +179,7 @@ func (tq *TransactionQuery) randOne(ctx context.Context) (string, error) {
 	}
 }
 
-// FromAny query transaction from any sharder that is not selected in previous queires. use any used sharder if there is not any unused sharder
-func (tq *TransactionQuery) FromAny(ctx context.Context, query string) (QueryResult, error) {
-
-	res := QueryResult{
-		StatusCode: http.StatusBadRequest,
-	}
-
-	err := tq.validate(1)
-
-	if err != nil {
-		return res, err
-	}
-
-	host, err := tq.randOne(ctx)
-
-	if err != nil {
-		return res, err
-	}
-
-	r := resty.New()
-	requestUrl := tq.buildUrl(host, query)
-
-	Logger.Debug("GET", requestUrl)
-
-	r.DoGet(ctx, requestUrl).
-		Then(func(req *http.Request, resp *http.Response, respBody []byte, cf context.CancelFunc, err error) error {
-			res.Error = err
-			if err != nil {
-				return err
-			}
-
-			res.Content = respBody
-			Logger.Debug(string(respBody))
-
-			if resp != nil {
-				res.StatusCode = resp.StatusCode
-			}
-
-			return nil
-		})
-
-	errs := r.Wait()
-
-	if len(errs) > 0 {
-		return res, errs[0]
-	}
-
-	return res, nil
-
-}
-
-// FromAll query transaction from all sharders whatever it is selected or offline in previous queires, and return consensus result
-func (tq *TransactionQuery) FromAll(ctx context.Context, query string, handle QueryResultHandle) error {
-	if tq == nil || tq.max == 0 {
-		return ErrNoAvailableSharders
-	}
-
-	urls := make([]string, 0, tq.max)
-	for _, host := range tq.sharders {
-		urls = append(urls, tq.buildUrl(host, query))
-	}
-
-	r := resty.New()
-	r.DoGet(ctx, urls...).
-		Then(func(req *http.Request, resp *http.Response, respBody []byte, cf context.CancelFunc, err error) error {
-			res := QueryResult{
-				Content:    respBody,
-				Error:      err,
-				StatusCode: http.StatusBadRequest,
-			}
-
-			if resp != nil {
-				res.StatusCode = resp.StatusCode
-
-				Logger.Debug(req.URL.String() + " " + resp.Status)
-				Logger.Debug(string(respBody))
-			} else {
-				Logger.Debug(req.URL.String())
-
-			}
-
-			if handle != nil {
-				if handle(res) {
-
-					cf()
-				}
-			}
-
-			return nil
-		})
-
-	r.Wait()
-
-	return nil
-}
-
-// GetFastConfirmation get txn confirmation from a random online sharder
-func (tq *TransactionQuery) GetFastConfirmation(ctx context.Context, txnHash string) (*blockHeader, map[string]json.RawMessage, *blockHeader, error) {
-	var confirmationBlockHeader *blockHeader
-	var confirmationBlock map[string]json.RawMessage
-	var lfbBlockHeader blockHeader
-
-	// {host}/v1/transaction/get/confirmation?hash={txnHash}&content=lfb
-	result, err := tq.FromAny(ctx, tq.buildUrl("", TXN_VERIFY_URL, txnHash, "&content=lfb"))
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	if result.StatusCode == http.StatusOK {
-
-		err = json.Unmarshal(result.Content, &confirmationBlock)
-		if err != nil {
-			Logger.Error("txn confirmation parse error", err)
-			return nil, nil, nil, err
-		}
-
-		// parse `confirmation` section as block header
-		confirmationBlockHeader, err = getBlockHeaderFromTransactionConfirmation(txnHash, confirmationBlock)
-		if err == nil {
-			return confirmationBlockHeader, confirmationBlock, nil, nil
-		}
-
-		Logger.Error("txn confirmation parse header error", err)
-
-		// parse `latest_finalized_block` section
-		lfbRaw, ok := confirmationBlock["latest_finalized_block"]
-		if !ok {
-			return confirmationBlockHeader, confirmationBlock, nil, err
-		}
-
-		err = json.Unmarshal([]byte(lfbRaw), &lfbBlockHeader)
-		if err == nil {
-			return confirmationBlockHeader, confirmationBlock, &lfbBlockHeader, ErrTransactionNotConfirmed
-		}
-
-		Logger.Error("round info parse error.", err)
-		return nil, nil, nil, err
-
-	}
-
-	return nil, nil, nil, thrown.Throw(ErrTransactionNotFound, strconv.Itoa(result.StatusCode))
-}
-
-func (tq *TransactionQuery) GetConsensusConfirmation(ctx context.Context, numSharders int, txnHash string) (*blockHeader, map[string]json.RawMessage, *blockHeader, error) {
-
+func (tq *TransactionQuery) getConsensusConfirmation(ctx context.Context, numSharders int, txnHash string) (*blockHeader, map[string]json.RawMessage, *blockHeader, error) {
 	maxConfirmation := int(0)
 	txnConfirmations := make(map[string]int)
 	var confirmationBlockHeader *blockHeader
@@ -404,45 +260,49 @@ func (tq *TransactionQuery) GetConsensusConfirmation(ctx context.Context, numSha
 	return confirmationBlockHeader, confirmationBlock, lfbBlockHeader, nil
 }
 
-func (tq *TransactionQuery) GetInfo(ctx context.Context, query string) (*QueryResult, error) {
+// getFastConfirmation get txn confirmation from a random online sharder
+func (tq *TransactionQuery) getFastConfirmation(ctx context.Context, txnHash string) (*blockHeader, map[string]json.RawMessage, *blockHeader, error) {
+	var confirmationBlockHeader *blockHeader
+	var confirmationBlock map[string]json.RawMessage
+	var lfbBlockHeader blockHeader
 
-	consensuses := make(map[int]int)
-	var maxConsensus int
-	var consensusesResp QueryResult
-	// {host}{query}
-	err := tq.FromAll(ctx, query,
-		func(qr QueryResult) bool {
-			//ignore response if it is network error
-			if qr.StatusCode >= 500 {
-				return false
-			}
-
-			consensuses[qr.StatusCode]++
-			if consensuses[qr.StatusCode] >= maxConsensus {
-				maxConsensus = consensuses[qr.StatusCode]
-				consensusesResp = qr
-			}
-
-			return false
-
-		})
-
+	// {host}/v1/transaction/get/confirmation?hash={txnHash}&content=lfb
+	result, err := tq.FromAny(ctx, tq.buildUrl("", TXN_VERIFY_URL, txnHash, "&content=lfb"))
 	if err != nil {
-		return nil, err
+		return nil, nil, nil, err
 	}
 
-	if maxConsensus == 0 {
-		return nil, errors.New("zcn: query not found")
+	if result.StatusCode == http.StatusOK {
+
+		err = json.Unmarshal(result.Content, &confirmationBlock)
+		if err != nil {
+			Logger.Error("txn confirmation parse error", err)
+			return nil, nil, nil, err
+		}
+
+		// parse `confirmation` section as block header
+		confirmationBlockHeader, err = getBlockHeaderFromTransactionConfirmation(txnHash, confirmationBlock)
+		if err == nil {
+			return confirmationBlockHeader, confirmationBlock, nil, nil
+		}
+
+		Logger.Error("txn confirmation parse header error", err)
+
+		// parse `latest_finalized_block` section
+		lfbRaw, ok := confirmationBlock["latest_finalized_block"]
+		if !ok {
+			return confirmationBlockHeader, confirmationBlock, nil, err
+		}
+
+		err = json.Unmarshal([]byte(lfbRaw), &lfbBlockHeader)
+		if err == nil {
+			return confirmationBlockHeader, confirmationBlock, &lfbBlockHeader, ErrTransactionNotConfirmed
+		}
+
+		Logger.Error("round info parse error.", err)
+		return nil, nil, nil, err
+
 	}
 
-	rate := float32(maxConsensus*100) / float32(tq.max)
-	if rate < consensusThresh {
-		return nil, ErrInvalidConsensus
-	}
-
-	if consensusesResp.StatusCode != http.StatusOK {
-		return nil, errors.New(string(consensusesResp.Content))
-	}
-
-	return &consensusesResp, nil
+	return nil, nil, nil, thrown.Throw(ErrTransactionNotFound, strconv.Itoa(result.StatusCode))
 }
